Fix dev env. setup script.
Modify setDevLinks.sh to make all the config and firmware
links needed to run NDDS simulation from your git directory.
1. Clone the NDDS project to /home/picarro/git (or clone it
somewhere else and soft link to /home/picarro/git.
2. cd to /home/picarro/git/host/Config/NDDS/bin
3. Run setDevLinks.sh
4. Run launchDevSimulatio.sh and the NDDS sim should run.

+++ Config/NDDS/bin/setDevLinks.sh
@@ -14,14 +14,75 @@
 # the processes can find their ini files, schemes,
 # and scripts.
 #
+# Edit:
+# Some config files now reference other files in the
+# binary installation directory '/home/picarro/I2000'.
+# Add additional soft links so that the files can
+# still be found without requiring us to build and
+# install a package.
+#
+# We are assuming this script is run from 
+# /home/picarro/git/host/Config/<analyzer type>/bin
+#
+cd ..
 cwd=$(pwd)
 
-rm ../../src/main/python/AppConfig
-rm ../../src/main/python/CommonConfig
-rm ../../src/main/python/InstrConfig
-rm ./CommonConfig
+# Set source tree links
+#
+if [ -d "../../src/main/python/AppConfig" ] 
+then
+    rm ../../src/main/python/AppConfig
+fi
 
-ln -s $cwd/AppConfig ../../src/main/python/AppConfig
-ln -s $cwd/../CommonConfig/ ../../src/main/python/
+if [ -d "../../src/main/python/CommonConfig" ] 
+then
+    rm ../../src/main/python/CommonConfig
+fi
+
+if [ -d "../../src/main/python/InstrConfig" ] 
+then
+    rm ../../src/main/python/InstrConfig
+fi
+
+if [ -d "./CommonConfig" ] 
+then
+    rm ./CommonConfig
+fi
+
 ln -s ../CommonConfig
-ln -s $cwd/InstrConfig/ ../../src/main/python/
+ln -s $cwd/AppConfig ../../src/main/python
+ln -s $cwd/CommonConfig/ ../../src/main/python
+ln -s $cwd/InstrConfig/ ../../src/main/python
+
+
+# Set installed package links
+#
+if [ -d "/home/picarro/I2000/AppConfig" ] 
+then
+    rm /home/picarro/I2000/AppConfig
+fi
+
+if [ -d "/home/picarro/I2000/CommonConfig" ] 
+then
+    rm /home/picarro/I2000/CommonConfig
+fi
+
+if [ -d "/home/picarro/I2000/InstrConfig" ] 
+then
+    rm /home/picarro/I2000/InstrConfig
+fi
+
+if [ -d "/home/picarro/I2000/Firmware" ] 
+then
+    rm /home/picarro/I2000/Firmware
+fi
+
+mkdir -p /home/picarro/I2000
+
+ln -s $cwd/AppConfig /home/picarro/I2000
+ln -s $cwd/CommonConfig /home/picarro/I2000
+ln -s $cwd/InstrConfig /home/picarro/I2000
+ln -s /home/picarro/git/host/src/main/python/Firmware /home/picarro/I2000
+
+cd /home/picarro/git/host/src/main/python/Host/Fitter
+bash makeFitutils
